Simple grammar improvements for Oathkeeper intro
Mostly replaces passive voice (which is a no-no in English) with active voice. I also added mention of attribute-based access control as well as usage of some of its terminology (policy enforcement point, policy decision point). I found this model to be one of the more intuitive ones. Feel free to disagree on this or anything else, then I'll take these parts out.