fix(docs): fix links to `toSession`
On the page https://www.ory.sh/docs/identities/session-to-jwt-cors/, right after the sentence `Great! Everything is set up! Let's convert an Ory Session to a JWT:`, when we click on the `REST API` tab, the link `For more details see the [API documentation](https://www.ory.sh/docs/identities/reference/api#tag/frontend/operation/toSession).` leads to a 404.

This PR fixes it and makes all links to `toSession` relative per the guidelines in the README.

Note 1: There are still ~200 or so links referring directly to `ory.sh/docs` instead of being relative. But this is a big task to update them all so I did not.

Note 2: `npm build` does detection of broken links but it did not catch this issue because the link actually points to an [existing file](https://github.com/ory/docs/blob/cf3c62ccc66dc4eb417c4aed13823778284c1949/docs/reference/api.mdx#L1-L1), just not the right one.