Remove unsupported settings hook
There is no sign of support for doing what was described in the documentation. The config schema suggests only "webhook" is allowed in definitions/selfServiceAfterSettingsMethod. In Kratos, I also do not see any sign of a ExecuteSettingsPostPersistHook method in address_verifier.go so do not understand how this could have ever worked.